Added getter methods on Appointment class,and in AssignmentOne class i have added ArrayList, cancelBooking ,printexisitingAppointments.
Here i have added getter methods on Appointment class.in AppointmentOne class i added ArrayList and create Appointment method to add new appointment  to the list which prints out all the information provided. pritnExistingAppointments method was added to display all the current Appointments if not appointment is pritns a message no appoitnmetn. canclebooking method was also added to cacncle appointment based on the patient phone number which prints error if number dont match, iteration is used to display it.so 2 each type of appoitnment is made and then cancel one appointment to to show the updated booking.